CR_12

+++ tokyo-supermix-server/src/main/java/com/tokyo/supermix/server/controller/TestParameterController.java
@@ -30,6 +30,7 @@ import com.tokyo.supermix.rest.enums.RestApiResponseStatus;
 import com.tokyo.supermix.rest.response.BasicResponse;
 import com.tokyo.supermix.rest.response.ContentResponse;
 import com.tokyo.supermix.rest.response.ValidationFailureResponse;
+import com.tokyo.supermix.server.services.TestConfigureService;
 import com.tokyo.supermix.server.services.TestParameterService;
 import com.tokyo.supermix.util.Constants;
 import com.tokyo.supermix.util.ValidationFailureStatusCodes;
@@ -41,6 +42,8 @@ public class TestParameterController {
   private TestParameterService testParameterService;
   @Autowired
   private ValidationFailureStatusCodes validationFailureStatusCodes;
+  @Autowired
+  private TestConfigureService testConfigureService;
   @Autowired
   private Mapper mapper;
   private static final Logger logger = Logger.getLogger(TestParameterController.class);
@@ -118,7 +121,14 @@ public class TestParameterController {
 
   @DeleteMapping(EndpointURI.TEST_PARAMETER_BY_ID)
   public ResponseEntity<Object> deleteTestParameter(@PathVariable Long id) {
+    TestParameter testParameter = testParameterService.getTestParameterById(id);
     if (testParameterService.isTestParameterExist(id)) {
+      if (testConfigureService.isAlreadyDepended(testParameter.getTestConfigure().getId())) {
+        return new ResponseEntity<>(
+            new ValidationFailureResponse(Constants.TEST_CONFIGURE,
+                validationFailureStatusCodes.getTestConfigureAlreadyDepended()),
+            HttpStatus.BAD_REQUEST);
+      }
       testParameterService.deleteTestParameter(id);
       return new ResponseEntity<>(
           new BasicResponse<>(RestApiResponseStatus.OK, Constants.TEST_PARAMETER_DELETED),
@@ -146,6 +156,12 @@ public class TestParameterController {
   public ResponseEntity<Object> updateTestParameter(
       @RequestBody TestParameterRequestDto testParameterRequestDto) {
     if (testParameterService.isTestParameterExist(testParameterRequestDto.getId())) {
+      if (testConfigureService.isAlreadyDepended(testParameterRequestDto.getTestConfigureId())) {
+        return new ResponseEntity<>(
+            new ValidationFailureResponse(Constants.TEST_CONFIGURE,
+                validationFailureStatusCodes.getTestConfigureAlreadyDepended()),
+            HttpStatus.BAD_REQUEST);
+      }
       if (testParameterService.isAbbreviationNull(testParameterRequestDto.getAbbreviation())) {
         return new ResponseEntity<>(new ValidationFailureResponse(Constants.ABBREVIATION,
             validationFailureStatusCodes.getAbbreviationIsNull()), HttpStatus.BAD_REQUEST);
